requirements.txt
Django==6.0.4
djangorestframework==3.17.1
xhtml2pdf==0.2.17
Pillow==12.2.0
reportlab==4.4.10
